+++ jaxkern/computations/__init__.py
@@ -1,0 +1,13 @@
+from .base import AbstractKernelComputation
+from .constant_diagonal import ConstantDiagonalKernelComputation
+from .dense import DenseKernelComputation
+from .diagonal import DiagonalKernelComputation
+from .eigen import EigenKernelComputation
+
+__all__ = [
+    "AbstractKernelComputation",
+    "ConstantDiagonalKernelComputation",
+    "DenseKernelComputation",
+    "DiagonalKernelComputation",
+    "EigenKernelComputation",
+]
