+++ jaxkern/stationary/matern12.py
@@ -1,0 +1,56 @@
+from typing import Dict, List, Optional
+
+import jax.numpy as jnp
+from jax.random import KeyArray
+from jaxtyping import Array, Float
+
+from ..base import AbstractKernel
+from ..computations import (
+    AbstractKernelComputation,
+    DenseKernelComputation,
+)
+from .utils import euclidean_distance
+
+
+class Matern12(AbstractKernel):
+    """The Matérn kernel with smoothness parameter fixed at 0.5."""
+
+    def __init__(
+        self,
+        compute_engine: AbstractKernelComputation = DenseKernelComputation,
+        active_dims: Optional[List[int]] = None,
+        stationary: Optional[bool] = False,
+        spectral: Optional[bool] = False,
+        name: Optional[str] = "Matérn 1/2 kernel",
+    ) -> None:
+        super().__init__(compute_engine, active_dims, stationary, spectral, name)
+
+    def __call__(
+        self,
+        params: Dict,
+        x: Float[Array, "1 D"],
+        y: Float[Array, "1 D"],
+    ) -> Float[Array, "1"]:
+        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with
+        lengthscale parameter :math:`\\ell` and variance :math:`\\sigma^2`
+
+        .. math::
+            k(x, y) = \\sigma^2 \\exp \\Bigg( -\\frac{\\lvert x-y \\rvert}{2\\ell^2}  \\Bigg)
+
+        Args:
+            params (Dict): Parameter set for which the kernel should be evaluated on.
+            x (Float[Array, "1 D"]): The left hand argument of the kernel function's call.
+            y (Float[Array, "1 D"]): The right hand argument of the kernel function's call
+        Returns:
+            Float[Array, "1"]: The value of :math:`k(x, y)`
+        """
+        x = self.slice_input(x) / params["lengthscale"]
+        y = self.slice_input(y) / params["lengthscale"]
+        K = params["variance"] * jnp.exp(-euclidean_distance(x, y))
+        return K.squeeze()
+
+    def init_params(self, key: KeyArray) -> Dict:
+        return {
+            "lengthscale": jnp.array([1.0] * self.ndims),
+            "variance": jnp.array([1.0]),
+        }
