+++ jaxkern/stationary/matern52.py
@@ -1,0 +1,59 @@
+from typing import Dict, List, Optional
+
+import jax.numpy as jnp
+from jax.random import KeyArray
+from jaxtyping import Array, Float
+
+from ..base import AbstractKernel
+from ..computations import (
+    AbstractKernelComputation,
+    DenseKernelComputation,
+)
+from .utils import euclidean_distance
+
+
+class Matern52(AbstractKernel):
+    """The Matérn kernel with smoothness parameter fixed at 2.5."""
+
+    def __init__(
+        self,
+        compute_engine: AbstractKernelComputation = DenseKernelComputation,
+        active_dims: Optional[List[int]] = None,
+        stationary: Optional[bool] = False,
+        spectral: Optional[bool] = False,
+        name: Optional[str] = "Matern 5/2",
+    ) -> None:
+        super().__init__(compute_engine, active_dims, stationary, spectral, name)
+
+    def __call__(
+        self, params: Dict, x: Float[Array, "1 D"], y: Float[Array, "1 D"]
+    ) -> Float[Array, "1"]:
+        """Evaluate the kernel on a pair of inputs :math:`(x, y)` with
+        lengthscale parameter :math:`\\ell` and variance :math:`\\sigma^2`
+
+        .. math::
+            k(x, y) = \\sigma^2 \\exp \\Bigg(1+ \\frac{\\sqrt{5}\\lvert x-y \\rvert}{\\ell^2} + \\frac{5\\lvert x - y \\rvert^2}{3\\ell^2} \\Bigg)\\exp\\Bigg(-\\frac{\\sqrt{5}\\lvert x-y\\rvert}{\\ell^2} \\Bigg)
+
+        Args:
+            params (Dict): Parameter set for which the kernel should be evaluated on.
+            x (Float[Array, "1 D"]): The left hand argument of the kernel function's call.
+            y (Float[Array, "1 D"]): The right hand argument of the kernel function's call.
+
+        Returns:
+            Float[Array, "1"]: The value of :math:`k(x, y)`.
+        """
+        x = self.slice_input(x) / params["lengthscale"]
+        y = self.slice_input(y) / params["lengthscale"]
+        tau = euclidean_distance(x, y)
+        K = (
+            params["variance"]
+            * (1.0 + jnp.sqrt(5.0) * tau + 5.0 / 3.0 * jnp.square(tau))
+            * jnp.exp(-jnp.sqrt(5.0) * tau)
+        )
+        return K.squeeze()
+
+    def init_params(self, key: KeyArray) -> Dict:
+        return {
+            "lengthscale": jnp.array([1.0] * self.ndims),
+            "variance": jnp.array([1.0]),
+        }
