+++ setup.cfg
@@ -6,6 +6,11 @@ versionfile_build = jaxkern/_version.py
 tag_prefix = v
 
 [flake8]
+max-line-length = 88
+max-complexity = 12
+ignore = E501
+select = C,E,F,W,B,B9
+extend-ignore = E203, W503, F722, F821
 exclude =
     versioneer.py
     jaxkern/_version.py
